Camera: Adding SD card removal event and send a stopPlayback call whenever such event occurs.
Whenever the SD card was removed during Playback, there was no
deinitialization. After a few iterations, this caused the video decoder
component to run out of memory eventually and not able to render video
anymore. The MovieView application Activity will handle SD card events
and if removed innapropriately, a stopPlayback will be called and the
video decoder component will be deinitialized.

CRs-Fixed: 211663

Conflicts:

	src/com/android/camera/MovieView.java